pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flowcept"
dynamic = ["version"]
requires-python = ">=3.10"

authors = [{name = "Oak Ridge National Laboratory"}]
description = "Capture and query workflow provenance data using data observability"
readme = "README.md"
license = "MIT"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3"
]
keywords = [
    "ai",
    "ml",
    "machine-learning",
    "provenance",
    "lineage",
    "responsible-ai",
    "databases",
    "big-data",
    "provenance",
    "tensorboard",
    "data-integration",
    "scientific-workflows",
    "dask",
    "reproducibility",
    "workflows",
    "parallel-processing",
    "lineage",
    "model-management",
    "mlflow",
    "responsible-ai",
    "data-analytics"
]

dependencies = [
    "omegaconf",
    "numpy",
    "msgpack",
    "orjson"
]

[project.optional-dependencies]
redis = ["redis"]
lmdb = ["lmdb"]
telemetry = ["psutil>=6.1.1", "py-cpuinfo"]
extras = ["flowcept[redis]", "flowcept[telemetry]", "flowcept[lmdb]", "GitPython", "pandas", "flask-restful", "requests"]

analytics = ["seaborn", "plotly", "scipy"]
mongo = ["pymongo", "pyarrow"]
dask = ["tomli", "dask[distributed]<=2024.10.0"]
docs = ["sphinx", "furo"]
kafka = ["confluent-kafka<=2.8.0"]  # As of today, 2/28/2025, version 2.8.1 is stale. When this gets fixed, let's remove the version constraint. https://pypi.org/project/confluent-kafka/#history
mlflow = ["mlflow-skinny", "SQLAlchemy", "alembic", "watchdog", "cryptography"]
nvidia = ["nvidia-ml-py"]
mqtt = ["paho-mqtt"]
tensorboard = ["tensorboard", "tensorflow", "tbparse"]
llm_agent = ["mcp[cli]", "langchain_community", "streamlit"]
llm_google = ["flowcept[llm_agent]", "google-genai"]

dev = [
    "flowcept[docs]",
    "jupyterlab",
    "nbmake",
    "pika",
    "pytest",
    "ruff",
    "pyyaml",
]
# Torch and some other ml-specific libs, only used for dev purposes, require the following specific versions.
ml_dev = [
    "torch==2.2.2",
    "torchvision==0.17.2",
    "torchtext==0.17.2",
    "datasets==2.17.0",
    "numpy<2.0",
    "sacremoses",
    "nltk"
]
all = [
    "flowcept[telemetry]",
    "flowcept[extras]",
    "flowcept[redis]",
    "flowcept[lmdb]",
    "flowcept[mongo]",
    "flowcept[analytics]",
    "flowcept[dask]",
    "flowcept[kafka]",
    "flowcept[mlflow]",
    "flowcept[mqtt]",
    "flowcept[tensorboard]",
    "flowcept[dev]",
    "flowcept[llm_agent]",
    "flowcept[llm_google]"
]

[tool.hatch.version]
path = "src/flowcept/version.py"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
extend-select = ["E501", "D"]
ignore = ["D200", "D212", "D105", "D401", "D205", "D100"]

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.hatch.build.targets.wheel]
packages = ["src/flowcept"]

[tool.hatch.build.targets.wheel.force-include]
"resources/sample_settings.yaml" = "resources/sample_settings.yaml"

[project.scripts]
flowcept = "flowcept.cli:main"